+++ handlers/common/redirect_test.go
@@ -1,4 +1,4 @@
-package handlers
+package common
 
 import (
 	"net/http"
