+++ handlers/common/static.go
@@ -1,4 +1,4 @@
-package handlers
+package common
 
 import (
 	"bytes"
